+++ src/ui/palette.rs
@@ -10,6 +10,7 @@ pub struct PaletteView {
     colors: Vec<IciColor>,
     selected: u8,
     cols: usize,
+    offset: isize,
 }
 
 impl PaletteView {
@@ -19,6 +20,7 @@ impl PaletteView {
             colors: vec![IciColor::transparent()],
             selected: 0,
             cols: 0,
+            offset: 0,
         }
     }
 }
@@ -26,7 +28,7 @@ impl PaletteView {
 impl PaletteView {
     pub fn set_palette(&mut self, new_colors: &[IciColor]) {
         self.colors = new_colors.to_vec();
-        self.cols = self.bounds.width() / PER_SQUARE;
+        self.cols = (self.bounds.width() / PER_SQUARE) + 1;
     }
 
     pub fn set_color_index(&mut self, idx: u8) {
@@ -41,17 +43,19 @@ impl PaletteView {
         if self.bounds.contains(mouse_xy) {
             let xy = mouse_xy - self.bounds.top_left();
             let x = xy.x / PER_SQUARE as isize;
-            let y = xy.y / PER_SQUARE as isize;
-            //the + y shouldn't be necessary
-            //this bug is most like due a difference between rendering and math
-            //the line is ending one too early when rendering or something
-            let i = x + y * (self.cols as isize) + y;
+            let y = (xy.y + self.offset) / PER_SQUARE as isize;
+            let i = x + y * (self.cols as isize);
             if i >= 0 && i < self.colors.len() as isize {
                 self.selected = i as u8;
                 return true;
             }
         }
         false
+    }
+
+    pub fn on_scroll(&mut self, _xy: Coord, y_diff: isize) {
+        self.offset += y_diff;
+        self.offset = self.offset.clamp(0, 100);
     }
 }
 
@@ -61,8 +65,8 @@ impl UiElement for PaletteView {
     }
 
     fn render(&self, graphics: &mut Graphics, _mouse_xy: Coord) {
-        let orig_trans = graphics.get_translate();
-        graphics.set_translate(self.bounds.top_left());
+        let orig_trans = graphics.set_translate(self.bounds.top_left() + (0, -self.offset));
+        graphics.clip_mut().set_valid_rect(self.bounds.clone());
 
         let mut x = 0;
         let mut y = 0;
@@ -88,13 +92,14 @@ impl UiElement for PaletteView {
                 );
             }
             x += 1;
-            if x > self.cols {
+            if x >= self.cols {
                 x = 0;
                 y += 1;
             }
         }
 
         graphics.set_translate(orig_trans);
+        graphics.clip_mut().set_all_valid();
     }
 
     fn update(&mut self, _: &Timing) {}